+++ src/main/kotlin/com/info/maeumgagym/security/authentication/token/vo/AuthenticationTokenType.kt
@@ -1,4 +1,4 @@
-package com.info.maeumgagym.security.jwt.vo
+package com.info.maeumgagym.security.authentication.token.vo
 
 /**
  * [AuthenticationToken] 및 그 외 임의의 토큰에서 타입을 표기하기 위한 enum
